Renames and updates import path for Astro component utility
Renames the Astro component utility file for improved clarity and updates its import path in the linting module. This change enhances code organization and readability.
