+++ src/lint/components/get-source-astro-component.ts
